repair seh expetion！

+++ src/PclSharp/yzsfilters.cpp
@@ -6,7 +6,7 @@
 #include "Helper.h"
 
 
-float* radius_down_sampling_run(const float* arrayF1, int rows, int* resultRow)
+float* radius_down_sampling_run(const float* arrayF1, int rows, size_t* resultRow)
 {
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud;
 	Helper::Create3DPointCloudFromArray<pcl::PointXYZ>(arrayF1, rows, cloud);
@@ -17,7 +17,7 @@ float* radius_down_sampling_run(const float* arrayF1, int rows, int* resultRow)
 	return floatArray;
 }
 
-float* statistical_outlier_filter(const float* array, const int rows, const int meanK, const float stddevMulThresh, int* resultRow)
+float* statistical_outlier_filter(const float* array, const int rows, const int meanK, const float stddevMulThresh, size_t* resultRow)
 {
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud;
 	Helper::Create3DPointCloudFromArray<pcl::PointXYZ>(array, rows, cloud);
@@ -26,9 +26,9 @@ float* statistical_outlier_filter(const float* array, const int rows, const int 
 	float* floatArray = Helper::PointCloudToArray(out_cloud);
 	*resultRow =out_cloud->points.size();
 	return floatArray;
-}
+}              
 
-float* smooth_filter(const float* array, const int rows,const int polynomialOrder, const float searchRadius, int* resultRow)
+float* smooth_filter(const float* array, const int rows,const int polynomialOrder, const float searchRadius, size_t* resultRow)
 {
 	pcl::PointCloud<pcl::PointXYZ>::Ptr cloud;
 	Helper::Create3DPointCloudFromArray<pcl::PointXYZ>(array, rows, cloud);
